// src/domain/enums/keyboard-layout-variant.ts
export enum KeyboardLayoutVariant {
  // English layouts
  QWERTY_US = "qwerty-us",
  QWERTY_UK = "qwerty-uk", 
  QWERTY_INTL = "qwerty-intl",
  DVORAK = "dvorak",
  COLEMAK = "colemak",
  
  // Lisu layouts
  SIL_BASIC = "sil-basic",
  SIL_STANDARD = "sil-standard",
  UNICODE_STANDARD = "unicode-standard",
  TRADITIONAL = "traditional",
  
  // Myanmar layouts
  MYANMAR3 = "myanmar3",
  ZAWGYI = "zawgyi",
  UNICODE_STANDARD_MY = "unicode-standard-my",
  WININNWA = "wininnwa"
}